Fix missing cast needed to lookup value correctly

+++ qpid-proton4j-engine/src/main/java/org/apache/qpid/proton4j/engine/impl/ProtonSessionIncomingWindow.java
@@ -159,7 +159,7 @@ public class ProtonSessionIncomingWindow {
             //        is navigable then we can use a sub-map to limit the range to the first
             //        and last elements and then simply walk next until the end without checking
             //        each index between for its presence.
-            ProtonIncomingDelivery delivery = unsettled.get(index);
+            ProtonIncomingDelivery delivery = unsettled.get((int) index);
             if (delivery != null) {
                 if (disposition.getSettled()) {
                     unsettled.remove((int) index);
